pyproject.toml
[project]
name = "ai-getter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.124.2",
    "google-genai>=1.54.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
